Add Wild schedule mock API
