+++ src/hooks/useWildSchedule.ts
@@ -1,15 +1,12 @@
-import useSWR from 'swr'
-
-export interface WildGame {
-  gameDate: string
-  opponent: string
-  home: boolean
-  [key: string]: any
-}
-
-const fetcher = (url: string) => fetch(url).then((res) => res.json())
+import { useState, useEffect } from 'react'
+import { getWildSchedule, type WildGame } from '@/lib/api'
 
 export default function useWildSchedule(limit = 1) {
-  const { data } = useSWR<WildGame[]>(`/api/nhl/schedule?limit=${limit}`, fetcher)
-  return data ?? null
+  const [data, setData] = useState<WildGame[] | null>(null)
+
+  useEffect(() => {
+    getWildSchedule(limit).then(setData)
+  }, [limit])
+
+  return data
 }
